backend/requirements.txt
albumentations==1.3.0
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
efficientnet-pytorch==0.7.1
fonttools==4.38.0
idna==3.4
imageio==2.22.4
inflate64==0.3.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
multivolumefile==0.2.3
munch==2.5.0
networkx==2.8.8
nibabel==4.0.2
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.3.0
pretrainedmodels==0.7.4
psutil==5.9.4
pybcj==1.0.1
pycryptodomex==3.15.0
pydicom==2.3.0
PyJWT==2.6.0
pyparsing==3.0.9
pyppmd==1.0.0
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
pyzstd==0.15.3
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
segmentation-models-pytorch==0.3.0
six==1.16.0
sqlparse==0.4.3
texttable==1.6.4
threadpoolctl==3.1.0
timm==0.4.12
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
